fix: colocate sidebar detail defaults helpers
